requirements.txt
fastapi[all]==0.103.2
python-dotenv==0.19.1
sqlalchemy==1.4.23
uvicorn==0.23.2
pytest==6.2.5
pydantic~=2.4.2